refactor: remove useless files and lib, change paths
