+++ package.json
@@ -25,17 +25,16 @@
     "@mui/material": "^5.15.1",
     "@reduxjs/toolkit": "^1.9.7",
     "@types/editorjs__header": "^2.6.3",
+    "@uiw/codemirror-theme-nord": "^4.21.21",
+    "@uiw/react-codemirror": "^4.21.21",
     "firebase": "^10.7.1",
     "formik": "^2.4.5",
     "notistack": "^3.0.1",
-    "@uiw/codemirror-theme-nord": "^4.21.21",
-    "@uiw/react-codemirror": "^4.21.21",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-hook-form": "^7.48.2",
     "react-redux": "^8.1.3",
     "react-router-dom": "^6.17.0",
-    "react-toastify": "^9.1.3",
     "yup": "^1.3.2",
     "yup-password": "^0.2.2"
   },
