+++ src/Components/VariablesEditor/VariablesEditor.tsx
@@ -2,7 +2,7 @@ import CodeEditor from '../CodeEditor/CodeEditor';
 import { FC } from 'react';
 import { json, jsonParseLinter } from '@codemirror/lang-json';
 import { linter } from '@codemirror/lint';
-import { useAppDispatch, useAppSelector } from '../../store/slices/hooks';
+import { useAppDispatch, useAppSelector } from '../../hooks/store';
 import { updateVariables } from '../../store/slices/querySlice';
 
 const VariablesEditor: FC = () => {
