--- src/pages/WelcomePage/developers.ts
@@ -1,26 +1,0 @@
-export const developers = [
-  {
-    name: 'Sergey',
-    bio: 'Est incididunt duis officia nulla officia in laboris officia voluptate veniam reprehenderit veniam enim ad. Dolore et sint adipisicing dolore aliqua ullamco. Laborum velit officia id dolor aliqua consequat.Anim do sit qui nulla commodo quis nostrud non esse Lorem. Est incididunt duis officia nulla officia in laboris officia voluptate veniam.',
-    location: 'Belarus, Minsk',
-    imgSrc: 'https://avatars.githubusercontent.com/u/52540855?v=4',
-    gitHub: 'https://github.com/surface74',
-    contribution: 'Frontend',
-  },
-  {
-    name: 'Elena',
-    bio: 'Anim do sit qui nulla commodo quis nostrud non esse Lorem. Est incididunt duis officia nulla officia in laboris officia voluptate veniam reprehenderit veniam enim ad. Dolore et sint adipisicing dolore aliqua ullamco. Laborum velit officia id dolor aliqua consequat.Anim do sit qui nulla commodo quis nostrud non esse Lorem. Est incididunt duis officia nulla officia in laboris officia voluptate veniam. Laborum incididunt duis officia commodo quis nostrud non esse Lorem. ',
-    location: 'Russia, Novosibirsk',
-    imgSrc: 'https://avatars.githubusercontent.com/u/87429007?v=4',
-    gitHub: 'https://github.com/MakaevaElena',
-    contribution: 'Frontend',
-  },
-  {
-    name: 'Valeriia',
-    bio: 'Anim do sit qui nulla commodo quis nostrud non esse Lorem. Est incididunt duis officia nulla officia in laboris officia voluptate veniam reprehenderit veniam enim ad. Dolore et sint adipisicing dolore aliqua ullamco. Laborum velit officia id dolor aliqua consequat.Anim do sit qui nulla commodo quis nostrud non esse Lorem. Est incididunt duis officia nulla officia in laboris officia voluptate veniam.',
-    location: 'Israel, Tel-Aviv',
-    imgSrc: 'https://avatars.githubusercontent.com/u/73602864?v=4',
-    gitHub: 'https://github.com/MaleryValery',
-    contribution: 'Frontend',
-  },
-];
